requirements.txt
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
ShopifyAPI==12.6.0
